// src/lit.smabu.react/src/types/domain/create-offer-command.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LIT.Smabu.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Currency } from './currency';
import { CustomerId } from './customer-id';
import { OfferId } from './offer-id';
import { OfferNumber } from './offer-number';
 /**
 * 
 *
 * @export
 * @interface CreateOfferCommand
 */
export interface CreateOfferCommand {

    /**
     * @type {OfferId}
     * @memberof CreateOfferCommand
     */
    id: OfferId;

    /**
     * @type {CustomerId}
     * @memberof CreateOfferCommand
     */
    customerId: CustomerId;

    /**
     * @type {Currency}
     * @memberof CreateOfferCommand
     */
    currency: Currency;

    /**
     * @type {number}
     * @memberof CreateOfferCommand
     */
    tax?: number;

    /**
     * @type {string}
     * @memberof CreateOfferCommand
     */
    taxDetails?: string | null;

    /**
     * @type {OfferNumber}
     * @memberof CreateOfferCommand
     */
    number?: OfferNumber;
}